Start moving Anki-specific files to a module
